refactor: imports and move layouts outside components dir
