+++ src/components/ImageCredits.tsx
