+++ src/components/ImageDimmer.tsx
