--- src/components/layouts/settings/SettingsTitle.tsx
@@ -1,9 +1,0 @@
-const SettingsTitle = ({ text = "Settings Title" }: { text?: string }) => {
-  return (
-    <h1 className="capitalize text-3xl font-bold dark:text-white w-max">
-      {text}
-    </h1>
-  );
-};
-
-export default SettingsTitle;
