--- src/components/layouts/settings/SettingsWrapper.tsx
@@ -1,5 +1,0 @@
-const SettingsWrapper = ({ children }: { children: React.ReactNode }) => {
-  return <div className="flex flex-col items-start gap-8">{children}</div>;
-};
-
-export default SettingsWrapper;
