--- src/components/layouts/settings/nav/SettingsNavSection.tsx
@@ -1,10 +1,0 @@
-const SettingsNavSection = ({ children }: { children: React.ReactNode }) => {
-  return (
-    <div className="flex flex-col space-y-2 mb-4">
-      {children}
-      <hr className="border-surface-dark dark:border-dm-surface " />
-    </div>
-  );
-};
-
-export default SettingsNavSection;
