--- src/components/sections/SettingSwitch.tsx
@@ -1,12 +1,0 @@
-type SettingSwitchProps = Partial<
-  Readonly<{
-    children: React.ReactNode;
-    className: string;
-  }>
->;
-
-const SettingSwitch = ({ children, className }: SettingSwitchProps) => {
-  return <div className={`w-max ${className}`}>{children}</div>;
-};
-
-export default SettingSwitch;
