--- src/constants/depts.ts
@@ -1,389 +1,0 @@
-import { deptsType } from "@src/shared/ts/type";
-
-export const deptsAnswer = "Here are the list of departments:";
-
-export const deptsMessages = [
-  "departments",
-  "department",
-  "can you give me the list of departments?",
-  "can you give me the list of departments",
-  "can you give me the departments?",
-  "can you give me the departments",
-  "can you give me list of departments?",
-  "can you give me list of departments",
-  "departments list?",
-  "departments list",
-  "department list?",
-  "department list",
-  "list of departments",
-  "list of department",
-  "give me the list of deparments",
-  "depts",
-  "do you have the departments list?",
-];
-
-export const depts: deptsType = [
-  {
-    id: "1",
-    deptName: "OFFICE OF THE CITY MAYOR (MO)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "2",
-    deptName: "OFFICE OF THE CITY VICE MAYOR (OVM)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "3",
-    deptName: "OFFICE OF THE CITY ADMINISTRATOR (ADMIN)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "4",
-    deptName: "BUSINESS PERMIT AND LICENSING DEPARTMENT (BPLO)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "5",
-    deptName: "OFFICE OF THE CITY BUILDING OFFICIAL (OBO)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "6",
-    deptName: "CITY PLANNING AND DEVELOPMENT COORDINATING OFFICE (CPDC)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "7",
-    deptName: "CITY ASSESSOR'S DEPARTMENT (ASSESSOR)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "8",
-    deptName: "BACOOR DISASTER RISK REDUCTION AND MANAGEMENT OFFICE (BDRRMO)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName:
-      "OFFICE OF THE ACCOUNTING AND INTERNAL AUDIT SERVICES (ACCOUNTING)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "ZONING AND LAND DEVELOPMENT DEPARTMENT (ZONING)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "CITY BACOOR TRAFFIC MANAGAMENT DEPARTMENT (BTMD)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "OFFICE OF THE SANGGUNIANG PANLUNGSOD (SP)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "CITY ENGINEERING OFFICE (ENGINEERING)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "OFFICE OF THE CITY CIVIL REGISTRY (CCR)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "OFFICE OF THE AGRICULTURAL SERVICES (AGRICULTURE)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "CITY ENVIRONMENT SERVICE DEPARTMENT (CESD)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "PUBLIC EMPLOYMENT SERVICE OFFICE (PESO)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "HUMAN RESOURCES DEVELOPMENT AND MANAGEMENT DEPARTMENT (HRDMD)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "OFFICE OF THE POPULATION DEVELOPMENT (POPULATION)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "OFFICE OF THE CITY LEGAL SERVICE (LEGAL)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "CITY BUDGET OFFICE (BUDGET)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "CITY FINANCE DEPARTMENT (CITY TREASURER'S OFFICE)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "OFFICE OF SOCIAL WELFARE AND DEVELOPMENT (CSWD)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "OFFICE OF THE CITY HEALTH SERVICES (CHO)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "OFFICE OF VETERINARY SERVICES (VET)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "CITY INFORMATION AND COMMUNITY RELATIONS DEPARTMENT (CICRD)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "OFFICE OF THE CITY GENERAL SERVICES (GSO)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "BIDS AND AWARDS COMMITTEE (BAC)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName:
-      "E-GOVERNANCE DEPARTMENT & MANAGEMENT INFORMATION SYSTEM (E-GOV/MIS)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "SPORTS DEVELOPMENT UNIT (SPORTS)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "BACOOR CITY CULTURE, HISTORY, ARTS AND TOURISM OFFICE",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "CITY LIVELIHOOD AND DEVELOPMENT DEPARTMENT (LIVELIHOOD)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "HOUSING, URBAN DEVELOPMENT AND RESETTLEMENT  DEPARTMENT (HUDRD)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "OFFICE FOR THE DEVELOPMENT OF COOPERATIVES (COOP)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "OFFICE OF THE SENIOR CITIZENS AFFAIRS (OSCA)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "DEPARTMENT OF PUBLIC SAFETY (BPSU)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName:
-      "LOCAL ECONOMIC DEVELOPMENT AND INVESTMENTS PROMOTIONS OFFICE (LEDIPO)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "ARTA",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "LOCAL YOUTH DEVELOPMENT OFFICE (LYDO)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-  {
-    id: "",
-    deptName: "CITY CEMETERY OFFICE (CEMETERY)",
-    service:
-      "* We can provide this \n* We can provide that \n* We can provide everything",
-    steps: "1. This is step one \n2. This is step two \n3. This is step three",
-    requirements:
-      "A. Bring this one \nB. Copy this to there \nC. Paste it here",
-  },
-];
