--- src/constants/settings/chat-head-size-list.ts
@@ -1,14 +1,0 @@
-export const chatHeadSizeList = [
-  {
-    name: "small",
-    value: 100,
-  },
-  {
-    name: "medium",
-    value: 200,
-  },
-  {
-    name: "large",
-    value: 300,
-  },
-];
