--- src/constants/settings/font-size-list.ts
@@ -1,26 +1,0 @@
-export const fontSizeList = [
-  {
-    name: "small",
-    value: 8,
-  },
-  {
-    name: "medium",
-    value: 12,
-  },
-  {
-    name: "normal",
-    value: 16,
-  },
-  {
-    name: "large",
-    value: 24,
-  },
-  {
-    name: "extra large",
-    value: 32,
-  },
-  {
-    name: "2XL",
-    value: 48,
-  },
-];
