--- src/constants/settings/settings-list.ts
@@ -1,15 +1,0 @@
-import { IoChatbubbleEllipsesOutline } from "react-icons/io5";
-import { FaVolumeHigh } from "react-icons/fa6";
-
-export const settingsList = [
-  {
-    name: "appearance",
-    settingType: "chat",
-    icon: IoChatbubbleEllipsesOutline,
-  },
-  {
-    name: "sound",
-    settingType: "chat",
-    icon: FaVolumeHigh,
-  },
-];
