+++ src/layouts/CityHallUI.tsx
@@ -1,11 +1,11 @@
 // Components
-import Section from "@components/pages/city-hall/common/Section";
+import Section from "@components/pages/city-hall/layouts/Section";
 import Header from "@components/pages/city-hall/layouts/header/Header";
 import TimeHeader from "@components/pages/city-hall/ui/TimeHeader";
 import Logo from "@components/pages/city-hall/ui/Logo";
 import ArticleLeft from "@components/pages/city-hall/articles/ArticleLeft";
 import ArticleRight from "@components/pages/city-hall/articles/ArticleRight";
-import Footer from "@src/components/pages/city-hall/layouts/footer/Footer";
+import Footer from "@components/pages/city-hall/layouts/footer/Footer";
 
 // Assets
 import LandingImg from "@static/assets/images/landing.png";
