+++ src/providers/ChatbotProvider.tsx
@@ -1,18 +1,14 @@
 import { useState, useEffect, useMemo } from "react";
+import { collection, doc, onSnapshot } from "firebase/firestore";
+
+// contexts
 import { ChatbotContext } from "@contexts/ChatbotContext";
 
-// db
-import { collection, doc, onSnapshot } from "firebase/firestore";
-import { db } from "../firebase/config";
-
 // types
-import {
-  ConfigurationType,
-  ConversationType,
-  FaqType,
-} from "@src/shared/ts/type";
+import { ConfigurationType, ConversationType, FaqType } from "@shared/ts/type";
 
-// lib
+// constants
+import { db } from "@constants/firebase/config";
 import { chatbotConfig } from "@constants/bot/chatbot-config";
 
 const botProfileCollectionRef = collection(db, "profile");
